+++ graphdatascience/tests/unit/test_dedicated_sessions.py
@@ -365,6 +365,56 @@ def test_create_attached_session(mocker: MockerFixture, aura_api: AuraApi) -> No
             uri="neo4j+s://foo.bar", username="client-id", password="client_secret"
         ),
         "session_id": "ffff0-ffff1",
+        "arrow_client_options": None,
+    }
+
+    assert isinstance(arrow_authentication, AuraApiTokenAuthentication)
+
+    assert len(sessions.list()) == 1
+    actual_session = sessions.list()[0]
+
+    assert actual_session.name == "my-session"
+    assert actual_session.user_id == "user-1"
+    assert actual_session.ttl == ttl
+
+
+def test_create_attached_session_passthrough_arrow_settings(mocker: MockerFixture, aura_api: AuraApi) -> None:
+    _setup_db_instance(aura_api)
+
+    sessions = DedicatedSessions(aura_api)
+
+    patch_construct_client(mocker)
+    patch_neo4j_query_runner(mocker)
+
+    ttl = timedelta(hours=42)
+    gds_parameters = sessions.get_or_create(
+        "my-session",
+        SessionMemory.m_8GB,
+        DbmsConnectionInfo("neo4j+s://ffff0.databases.neo4j.io", "dbuser", "db_pw"),
+        ttl=ttl,
+        arrow_client_options={"foo": "bar"},
+    )
+
+    arrow_authentication = gds_parameters["arrow_authentication"]  # type: ignore
+    del gds_parameters["arrow_authentication"]
+
+    dbms_authentication = gds_parameters["db_runner"].pop("auth")  # type: ignore
+
+    assert (dbms_authentication.principal, dbms_authentication.credentials) == ("dbuser", "db_pw")
+
+    assert gds_parameters == {  # type: ignore
+        "db_runner": {
+            "endpoint": "neo4j+s://ffff0.databases.neo4j.io",
+            "aura_ds": True,
+            "database": None,
+            "show_progress": False,
+            "config": None,
+        },
+        "session_bolt_connection_info": DbmsConnectionInfo(
+            uri="neo4j+s://foo.bar", username="client-id", password="client_secret"
+        ),
+        "session_id": "ffff0-ffff1",
+        "arrow_client_options": {"foo": "bar"},
     }
 
     assert isinstance(arrow_authentication, AuraApiTokenAuthentication)
@@ -412,6 +462,7 @@ def test_create_standalone_session(mocker: MockerFixture, aura_api: AuraApi) -> 
             uri="neo4j+s://foo.bar", username="client-id", password="client_secret"
         ),
         "session_id": "None-ffff0",
+        "arrow_client_options": None,
     }
 
     assert isinstance(arrow_authentication, AuraApiTokenAuthentication)
@@ -465,6 +516,7 @@ def test_get_or_create(mocker: MockerFixture, aura_api: AuraApi) -> None:
             uri="neo4j+s://foo.bar", username="client-id", password="client_secret"
         ),
         "session_id": "ffff0-ffff1",
+        "arrow_client_options": None,
     }
 
     assert gds_args1 == gds_args2